Auth/requirements.txt
asgiref==3.2.7
astroid==2.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.1
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.3
idna==2.9
importlib-metadata==1.6.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
mysqlclient==1.4.6
PyJWT==1.7.1
pylint==2.2.2
PyMySQL==0.9.3
pytz==2020.1
requests==2.23.0
six==1.12.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.1
zipp==3.1.0
